+++ src/message/groups/non_trans_receipt_couples.rs
@@ -60,7 +60,7 @@ impl GroupItem for NonTransReceiptCouple {
         self.cigar.qb2().map_err(ParsideError::from)
     }
 
-    fn full_size(&self) -> ParsideResult<u32> {
+    fn full_size(&self) -> ParsideResult<usize> {
         let size = self.cigar.full_size()?;
         Ok(size)
     }
