+++ src/message/groups/trans_last_idx_sig_groups.rs
@@ -65,40 +65,40 @@ impl TransLastIdxSigGroup {
 
 impl GroupItem for TransLastIdxSigGroup {
     fn qb64(&self) -> ParsideResult<String> {
-        let mut out = "\0".repeat(self.full_size()? as usize);
+        let mut out = "\0".repeat(self.full_size()?);
         let mut offset = 0;
-        let mut len = self.prefixer.full_size()? as usize;
+        let mut len = self.prefixer.full_size()?;
         unsafe { out[offset..len].as_bytes_mut() }
             .copy_from_slice(self.prefixer.qb64()?.as_bytes());
         offset += len;
-        len = self.isigers.full_size()? as usize;
+        len = self.isigers.full_size()?;
         unsafe { out[offset..len].as_bytes_mut() }.copy_from_slice(self.isigers.qb64()?.as_bytes());
         Ok(out)
     }
 
     fn qb64b(&self) -> ParsideResult<Vec<u8>> {
-        let mut out = vec![0u8; self.full_size()? as usize];
+        let mut out = vec![0u8; self.full_size()?];
         let mut offset = 0;
-        let mut len = self.prefixer.full_size()? as usize;
+        let mut len = self.prefixer.full_size()?;
         out[offset..len].copy_from_slice(&self.prefixer.qb64b()?);
         offset += len;
-        len = self.isigers.full_size()? as usize;
+        len = self.isigers.full_size()?;
         out[offset..len].copy_from_slice(&self.isigers.qb64b()?);
         Ok(out)
     }
 
     fn qb2(&self) -> ParsideResult<Vec<u8>> {
-        let mut out = vec![0u8; self.full_size()? as usize / 4 * 3];
+        let mut out = vec![0u8; self.full_size()? / 4 * 3];
         let mut offset = 0;
-        let mut len = self.prefixer.full_size()? as usize / 4 * 3;
+        let mut len = self.prefixer.full_size()? / 4 * 3;
         out[offset..len].copy_from_slice(&self.prefixer.qb2()?);
         offset += len;
-        len = self.isigers.full_size()? as usize / 4 * 3;
+        len = self.isigers.full_size()? / 4 * 3;
         out[offset..len].copy_from_slice(&self.isigers.qb2()?);
         Ok(out)
     }
 
-    fn full_size(&self) -> ParsideResult<u32> {
+    fn full_size(&self) -> ParsideResult<usize> {
         let size = self.prefixer.full_size()? + self.isigers.full_size()?;
         Ok(size)
     }
